requirements.txt
python-telegram-bot==21.6
fastapi==0.115.0
uvicorn==0.30.0
aiosqlite==0.20.0
python-dotenv==1.0.1
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
